quick fixes for compatibility and search problem
Line 714: The idea here is to only catch keys if they are a `:BIBTEX-KEY:` property or at the start of an org-mode headline. Someone familiar with using regexps in org-mode could do this much better.

Lines 716-718: I don't see `outline-hide-other` and `outline-show-subtree` defined in Emacs 24.3.1. I'm not sure what the side-effects of this change may be.