add fallback options to ivy-bibtex
This is done by introducing a new action offering to select a fallback
options (this is the easiest way I could think of, since ivy-read has no
builtin mechanism to simultaneously use different sources with different
matchers and actions).

To do this efficiently I made the search expression a variable of the
functions in bibtex-completion-fallback-options (helm-pattern or
ivy-text, depending on the backend). I updated the readme accordingly.