make ivy-bibtex-open-any the default action
To do so we need to bind `ivy-bibtex-default-action' to it rather than
putting it first in the action list (in ivy the action list is added
after the default action, and is automatically bound to `o'. Then we can
keep the old `ivy-bibtex-open-pdf', bound to `p' in the action list, so
both actions are available.