make candidate format customizable
Created a new custom variable `bibtex-completion-display-formats` and
modified the function `bibtex-completion-format-entry` accordingly. The
former variable is an alist of format strings for displaying entries in the results list. The key of each element of this list is either a BibTeX entry type (in which case the format string applies to entries of this type only) or t (in which case the format string applies to all other entry types). The value is the format string.

In the format string, expressions like \"${author}\", \"${title}\", etc,
are expanded to the value of the corresponding field. An expression like
\"${author:N}\" is further truncated to a width of N characters, whereas
an expression like \"${title:-N}\" is truncated to a width equal to the
results window's width minus N characters. Three special fields are
available: \"=type=\" holds the BibTeX entry type, \"=has-pdf=\" holds
`bibtex-completion-pdf-symbol' if the entry has a pdf file, and
\"=has-notes=\" holds `bibtex-completion-notes-symbol' if the entry has
a notes file.

The default format is unchanged.