Escape regexp chars to avoid creating malformed/mismatched regexps
Escape/quote special regexp characters in bibtex keys to avoid
inadvertently mis-translating the keys when creating this search regexp.

Typically this is not encountered since keys ideally omit punctuation,
but a wild "[" in a key yields an invalid regexp or a "*" likely
yields a non-matching regexp.  In practicality, this only likely to
occur with dirty data, but this change improves the durability of the code.