formatting references truncated the first author
In the function bibtex-completion-apa-format-authors, long author lists
were truncated to eight authors using the function -slice. However, the
call to slice used the hard-coded arguments 1 and 8 to delimit the first
and last authors to retain. This incorrect, as the first list element
is 0 in -slice (i.e., it is 0-indexed.)