Find PDFs of cross-referenced entries
Add optional argument FIND-CROSSREF to `bibtex-completion-find-pdf`
that appends PDFs from cross-referenced entries. This argument is set
to t in action functions (no need to do this when building the
bibliography because the "=has-pdf=" field is inherited from the
cross-referenced entry).

Also `bibtex-completion-find-pdf` now calls
`bibtex-completion-get-entry1` directly if needed and makes sure to
pass an entry to `bibtex-completion-find-pdf-in-field` or a key to
`bibtex-completion-find-pdf-in-library` (which allowed small
simplifications of these latter two functions). This avoids situations
where
`bibtex-completion-get-entry1` would be called twice for the same key.