Notes: narrow to the complete template instead of last org-heading
Before, a note template consisting of sub-headings would result in
only showing that sub-heading, e.g. using a template like this:

* Author (XXX): Title
** takeaways
** new questions
** additional literature

would only narrow to additional literature.  Now we narrow to the
complete template instead.

A side-effect is that for non-org templates the point is now at the
start of the template instead of the end.