Prevent squashing of newlines after template
The default behaviour of `org-capture-fill-template` is to squash all the
newlines at the end of the template into one.  This is problematic when your
template ends with multiple line.

By including non-newline characters at the end of the template,
`org-capture-fill-template` preserves those spaces.  The included non-newline
character can then be removed.

Definitely hackish.  An easier solution might be available.