Fix entries overflowing if they include wide characters
Using eg. "${title:*}" in bibtex-completion-display-formats should
truncate that field to the remaining window width, but previously that
calculation fails to account for characters wider than one column,
like CJK characters or tabstop.

This fixes that by using `string-width` (added in Emacs 20) instead of
`length` to determine the displayed width of the string.