Added bibtex-completion-get-entry-with-string.
Added bibtex-completion-find-key-from-file,
bibtex-completion-find-key-functions and
bibtex-completion-find-key-from-note.

The find-key-from-file runs bibtex-completion-find-key-functions to
return a key associated with the file name passed to it.

bibtex-completion-find-key-from-note is one of the
bibtex-completion-find-key-functions functions.

It uses a file name then strip it to its base. For example
"/path/to/author99.org" will be stripped to "author99" then—assuming
that "author99" is a key—it double-checks by passing "author99" to
bibtex-completion-find-note-functions, which returns a note file path
associated with it, only returning the key if the file are the same.

Will add find-key-from-pdf soon.