Avoid infinite loop when parsing a .bib file.
In parsebib.el 4.7 / 5.0, the function parsebib-find-next-item leaves point
after the @-sign when it finds an item. In parsebib.el 6.0, it leaves
point *before* the @-sign. Therefore, the caller must advance point if the
item is not read, otherwise parsebib-find-next-item will keep finding the
same item.

This fixes the problems reported in https://github.com/tmalsburg/helm-bibtex/issues/452 and
https://github.com/jkitchin/scimax/issues/510.